% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package-code.R
\name{run_test_dir}
\alias{run_test_dir}
\title{Simpulate test_check ortherwise running test_dir might skip some tests.
Based on testthat:::test_package_dir.}
\usage{
run_test_dir(package, path, ...)
}
\description{
Simpulate test_check ortherwise running test_dir might skip some tests.
Based on testthat:::test_package_dir.
}
