% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package-code.R
\name{run_test_env}
\alias{run_test_env}
\title{Create a package environment with access to all package functions.
Based on testthat:::test_pkg_env.}
\usage{
run_test_env(package)
}
\description{
Create a package environment with access to all package functions.
Based on testthat:::test_pkg_env.
}
