% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package-code.R
\name{run_all}
\alias{run_all}
\title{on purpose it only uses base functions}
\usage{
run_all(
  package,
  runnable_code_file,
  runnable_code_path = dirname(runnable_code_file),
  run_before = NULL,
  run_after = NULL
)
}
\description{
on purpose it only uses base functions
}
